:construction: 처음써보는 제너릭 :construction:

+++ app/src/main/kotlin/me/sungbin/spakchat/util/Util.kt
@@ -3,7 +3,7 @@ package me.sungbin.spakchat.util
 import com.sungbin.androidutils.util.Util
 
 object Util {
-    val randomId get() = Util.makeRandomUUID().replace("-", "").substring(0..8)
+    val randomId get() = Util.makeRandomUUID().replace("-", "").substring(0..8).toLong()
 
     fun generateMessageId(message: String, ownerName: String) =
         "${EncryptUtil.encrypt(EncryptUtil.EncryptType.MD5, message).substring(0..3)}${
@@ -11,5 +11,5 @@ object Util {
                 EncryptUtil.EncryptType.SHA256,
                 ownerName
             ).substring(0..3)
-        }$randomId".substring(0..3)
+        }$randomId".substring(0..3).toLong()
 }
